// src/app/api/model/chatUpdateRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Запрос для обновления информации о чате
 */
export interface ChatUpdateRequest { 
    /**
     * Новое название чата
     */
    name?: string;
    /**
     * Новое описание чата
     */
    description?: string;
}