// src/app/api/model/chatUserDeleteRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Запрос на удаление пользователей из чата
 */
export interface ChatUserDeleteRequest { 
    /**
     * Список идентификаторов пользователей для чата. Все ID должны быть положительными.
     */
    userIds?: Array<number>;
}