// src/app/api/model/userDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Представление данных пользователя, включая ID, имя, email и дату создания
 */
export interface UserDTO { 
    /**
     * Идентификатор пользователя
     */
    id: number;
    /**
     * Имя пользователя
     */
    username: string;
    /**
     * Адрес электронной почты
     */
    email: string;
    /**
     * Дата и время создания пользователя
     */
    createdAt: string;
}